Refactor dist and uncertain classes
dist and uncertain classes now all inherit from Sampleable. This means we can use Uncertain and distributions as inputs for other distributions.

Kind of uneeded but I want these objects to be treated like numbers as much as possible. Also allows x + beta(1,2) or somethiing in constructor which may be useful.